+++ App/view/error/404.php
@@ -2,6 +2,6 @@
     <h1><?php TP('page not found'); ?></h1>
     <br><br><br><br><br><br><br>
     <div>
-        <button class="ui button" onclick="toHome();"><?php TP('Return'); ?></button>
+        <button class="btn btn-default" onclick="toHome();"><?php TP('Return'); ?></button>
     </div>
 </div>