+++ sql/uninstall.sql
@@ -6,5 +6,8 @@ DROP FUNCTION IF EXISTS public.add_user_role(TEXT, BIGINT) CASCADE;
 DROP FUNCTION IF EXISTS public.delete_user_role(TEXT, BIGINT) CASCADE;
 DROP FUNCTION IF EXISTS public.active_user(TEXT) CASCADE;
 DROP FUNCTION IF EXISTS public.change_roles(TEXT, JSON) CASCADE;
+DROP FUNCTION IF EXISTS public.tp_change_department() CASCADE;
+DROP FUNCTION IF EXISTS public.tp_change_role() CASCADE;
+DROP FUNCTION IF EXISTS public.tp_change_user() CASCADE;
 DROP TABLE IF EXISTS public.user_log, public.user_role, public.users, public.roles, public.departments CASCADE;
 DROP SEQUENCE IF EXISTS public.user_id_seq;